Extract the things in common between phone, email, and address into parent class Contact after creating it.
